update commonapp/graphapp to use graphs only
