+++ pkg/Apps/CommonApp/arcs/services.js
@@ -16,7 +16,8 @@ const load = async paths => (await Promise.all(paths.map(p => import(`${Library}
 
 const modules = await load([
   'Mediapipe/PoseService.js',
-  'App/HistoryService.js'
+  'App/HistoryService.js',
+  'Arc/ArcService.js',
   // 'NewMedia/MediaService.js',
   // 'Goog/GoogleApisService.js',
   // 'Mediapipe/FaceMeshService.js',
