+++ pkg/Library/App/Graph/GraphApp.js
@@ -5,46 +5,37 @@
  * license that can be found in the LICENSE file.
  */
 import {App} from '../Worker/App.js';
-// import {RecipeBuilder} from '../../RecipeBuilder/RecipeBuilder.js';
+
+const GraphRecipe = {
+  $meta: {name: 'GraphRecipe', id: 'GraphRecipe'},
+  main: {
+    // n.b. the GraphRecipe actually only wants the rectangles from the graph layouts
+    $kind: '$app/Library/Graph'
+  }
+};
 
 export const GraphApp = class extends App {
   async spinup() {
-    this.layoutId ??= 'preview';
-    this.recipes = this.configureRecipes(this.graph, this.layoutId);
+    Object.assign(this.nodeTypes, {GraphRecipe});
+    this.graphs = [
+      this.graph,
+      this.configureGraph(this.graph)
+    ];
     this.logInfo();
     await super.spinup();
-    await this.addGraph(this.graph, this.layoutId);
   }
-  configureRecipes(graph, layoutId) {
-    // n.b. the GraphRecipe actually only wants the rectangles from the graph layouts
-    const GraphRecipe = {
-      main: {
-        $kind: '$app/Library/Graph',
-        $staticInputs: {graph, layoutId}
-      }
+
+  configureGraph(graph, layoutId) {
+    return {
+      $meta: {
+        id: 'graph-graph',
+        name: 'graph-graph'
+      },
+      nodes: [{
+        type: 'GraphRecipe',
+        props: {graph, layoutId}
+      }]
     };
-    return [GraphRecipe];
-  }
-  async onservice(user, host, request) {
-    if (request?.msg === 'addRunnerGraph') {
-      if (this.lastGraph) {
-        await this.removeGraph(this.lastGraph);
-      }
-      this.lastGraph = request.data.graph;
-      if (this.lastGraph) {
-        await this.addGraph(this.lastGraph);
-      }
-    }
-  }
-  async removeGraph(graph) {
-    this.arcs.removeGraph('user', graph, this.nodeTypes);
-  }
-  async addGraph(graph, layoutId) {
-    if (this.lastGraph) {
-      await this.removeGraph(this.lastGraph);
-      this.lastGraph = null;
-    }
-    await this.arcs.addGraph('user', graph, this.nodeTypes);
   }
 
   logInfo() {
