+++ pkg/Library/Designer/Designer.js
@@ -5,140 +5,22 @@
  * license that can be found in the LICENSE file.
  */
 ({
-idDelimiter: '$$',
-
-async initialize(inputs, state) {
-  state.recipes = {};
-  state.stores = {};
-},
-
-async update({recipes, selectedNodeId}, state, {service}) {
-  recipes ??= [];
-  // reset selection state
-  state.selectedNodeId = selectedNodeId;
-  await this.removeOldRecipes(recipes, state, service);
-  await this.renewRecipes(recipes, state, service);
-},
-
-async removeOldRecipes(recipes, state, service) {
-  return Promise.all(values(state.recipes).map(async recipe => {
-    if (!this.findRecipe(recipe.$meta.name, recipes)) {
-      return this.removeRecipe(recipe, state, service);
-    }
-  }));
-},
-
-async removeRecipe(recipe, state, service) {
-  await this.stopRecipe(recipe, state, service);
-  await this.removeStores(recipe, state, service);
-},
-
-findRecipe(recipeName, recipes) {
-  return recipes.find(({$meta: {name}}) => recipeName === name);
-},
-
-async stopRecipe(recipe, state, service) {
-  delete state.recipes[recipe.$meta.name];
-  return service({kind: 'RecipeService', msg: 'FinagleRecipe', data: {recipe, value: false}});
-},
-
-async removeStores({$stores}, state, service) {
-  return Promise.all(entries($stores || {}).map(async ([storeId, store]) => {
-    // if (!store.connection) {
-    delete state.stores[storeId];
-    await service({kind: 'StoreService', msg: 'RemoveStore', data: {storeId}});
-    // }
-  }));
-},
-
-async renewRecipes(recipes, state, service) {
-  for (const recipe of recipes) {
-    const runningRecipe = state.recipes[recipe.$meta.name];
-    if (!runningRecipe || runningRecipe.$meta.custom) {
-      if (runningRecipe && !deepEqual(runningRecipe, recipe)) {
-        await this.stopRecipe(recipe, state, service);
-      }
-      await this.startRecipe(recipe, state, service);
-    } else {
-      await this.updateStores(recipe, runningRecipe, state, service);
-      await this.updateConnections(recipe, runningRecipe, state, service);
-      await this.updateContainers(recipe, runningRecipe, state, service);
-    }
+async update({graph}, state, {service}) {
+  if (state.graph && graph?.$meta?.id !== state.graph?.$meta?.id) {
+    await service({kind: 'ArcService', msg: 'removeGraph', data: {graph: state.graph}});
   }
-},
-
-async updateStores(recipe, runningRecipe, state, service) {
-  const storeIds = this.findMissingStores(recipe, runningRecipe);
-  storeIds.forEach(storeId => {
-    service({kind: 'StoreService', msg: 'AddStore', data: {storeId, store: recipe.$stores[storeId]}});
-    state.recipes[recipe.$meta.name] = recipe;
-  });
-},
-
-async updateConnections(recipe, runningRecipe, state, service) {
-  const particles = this.findParticlesWithChangedConnections(recipe, runningRecipe);
-  particles.forEach(particleId => {
-    service({kind: 'RecipeService', msg: 'UpdateConnections', data: {particleId, spec: recipe[particleId]}});
-    state.recipes[recipe.$meta.name] = recipe;
-  });
-},
-
-async updateContainers(recipe, runningRecipe, state, service) {
-  const hostIds = this.findParticlesWithChangedContainer(recipe, runningRecipe);
-  if (hostIds.length > 0) {
-    const container = recipe[hostIds[0]].$container;
-    await service({kind: 'ComposerService', msg: 'setContainer', data: {hostIds, container}});
-    state.recipes[recipe.$meta.name] = recipe;
+  if (graph && !deepEqual(graph, state.graph)) {
+    // state.particleIds = 
+    await service({kind: 'ArcService', msg: 'runGraph', data: {graph}});
   }
+  state.graph = graph;
 },
 
-omitConnectionStores(recipe, state) {
-  entries(recipe.$stores || {}).forEach(([storeId, store]) => {
-    if (store.connection && state.stores[storeId]) {
-      delete recipe.$stores[storeId];
-    } else {
-      state.stores[storeId] = store;
-    }
-  });
-  return recipe;
-},
-
-async startRecipe(recipe, state, service) {
-  state.recipes[recipe.$meta.name] = recipe;
-  return service({kind: 'RecipeService', msg: 'FinagleRecipe', data: {
-    recipe: this.omitConnectionStores(recipe, state),
-    value: true
-  }});
-},
-
-findMissingStores(recipe, runningRecipe) {
-  return keys(recipe.$stores).filter(storeId => !runningRecipe.$stores[storeId]);
-},
-
-findParticlesWithChangedConnections(recipe, runningRecipe) {
-  const particleNames = this.getParticleNames(recipe);
-  return particleNames.filter(particleName => {
-    return JSON.stringify(recipe[particleName].$inputs) !== JSON.stringify(runningRecipe[particleName].$inputs);
-  });
-},
-
-findParticlesWithChangedContainer(recipe, runningRecipe) {
-  const particleNames = this.getParticleNames(recipe);
-  return particleNames.filter(particleName => {
-    return recipe[particleName].$container !== runningRecipe[particleName].$container;
-  });
-},
-
-getParticleNames(recipe) {
-  const notKeyword = name => !name.startsWith('$');
-  return recipe && keys(recipe).filter(notKeyword);
-},
-
-render({graph, nodeTypes, categories, layoutId}, {selectedNodeId, recipes}) {
+render({graph, nodeTypes, categories, layoutId}, {selectedNodeId}) {
   const node = graph?.nodes?.[selectedNodeId];
   const nodeType = nodeTypes?.[node?.type];
-  const ids = this.particleIdsForNode(node, graph, recipes);
-  const rects = this.getRects(graph, recipes, layoutId);
+  const ids = this.particleIdsForNode(node, graph, nodeTypes);
+  const rects = this.getRects(graph, nodeTypes,layoutId);
   return {
     rects,
     selectedKeys: ids,
@@ -146,24 +28,26 @@ render({graph, nodeTypes, categories, layoutId}, {selectedNodeId, recipes}) {
   };
 },
 
-particleIdsForNode(node, graph, recipes) {
+particleIdsForNode(node, graph, nodeTypes) {
   return node
     && !this.isUIHidden(node)
-    && this.getParticleNamesForNode(node, graph, recipes)
+    && this.getParticleNamesForNode(node, graph, nodeTypes)
     || []
     ;
 },
 
-getParticleNamesForNode(node, graph, recipes) {
+getParticleNamesForNode(node, graph, nodeTypes) {
   if (graph) {
-    const recipe = recipes[this.encodeFullNodeId(node, graph)];
-    return this.getParticleNames(recipe);
+    const nodeType = nodeTypes[node.type];
+    return keys(nodeType)
+      .filter(key => !key.startsWith('$'))
+      .map(key => `${node.id}:${key}`);
   }
 },
 
-getRects(graph, recipes, layoutId) {
+getRects(graph, nodeTypes, layoutId) {
   const rectMap = (id, node) => ({id, position: graph.layout?.[layoutId]?.[node.id]});
-  const nodeMap = node => this.particleIdsForNode(node, graph, recipes).map(id => rectMap(id, node));
+  const nodeMap = node => this.particleIdsForNode(node, graph, nodeTypes).map(id => rectMap(id, node));
   return values(graph?.nodes).map(nodeMap).flat();
 },
 
@@ -172,25 +56,27 @@ isUIHidden(node) {
 },
 
 onNodeDelete({eventlet: {key}, graph}, state) {
-  const node = this.findNodeByParticle(key, graph, state.recipes);
-  delete graph.nodes[node.id];
-  return {
-    graph,
-    ...this.selectNode(null, state)
-  };
+  // TODO(mariakleiner): update to not use RECIPES
+  // const node = this.findNodeByParticle(key, graph, state.recipes);
+  // delete graph.nodes[node.id];
+  // return {
+  //   graph,
+  //   ...this.selectNode(null, state)
+  // };
 },
 
 onNodePosition({eventlet: {key, value}, graph, layoutId}, state) {
-  const node = this.findNodeByParticle(key, graph, state.recipes);
-  if (!node) {
-    return this.selectNode(null, state);
-  }
-  ((graph.layout ??= {})[layoutId] ??= {})[node.id] = value;
-  //console.log('caching layout rect', node.id, value);
-  return {
-    graph,
-    ...this.selectNode(node.id, state)
-  };
+  // TODO(mariakleiner): update to not use RECIPES
+  // const node = this.findNodeByParticle(key, graph, state.recipes);
+  // if (!node) {
+  //   return this.selectNode(null, state);
+  // }
+  // ((graph.layout ??= {})[layoutId] ??= {})[node.id] = value;
+  // //console.log('caching layout rect', node.id, value);
+  // return {
+  //   graph,
+  //   ...this.selectNode(node.id, state)
+  // };
 },
 
 selectNode(id, state) {
